backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
anthropic==0.40.0
PyPDF2==3.0.1
python-dotenv==1.0.1
werkzeug==3.1.3
gunicorn==23.0.0
python-docx==1.1.2
